+++ contracts/transmuter/src/swap/mod.rs
@@ -16,1181 +16,23 @@ mod tests {
     use std::collections::BTreeMap;
 
     use crate::scope::Scope;
-    use crate::swap::common::construct_scope_value_pairs;
+
     use crate::{
         asset::Asset,
-        corruptable::Corruptable,
         swap::common::{Entrypoint, SwapExactAmountInResponseData, SwapExactAmountOutResponseData},
-        transmuter_pool::AssetGroup,
         ContractError,
     };
     use crate::{contract::Transmuter, transmuter_pool::TransmuterPool};
     use cosmwasm_std::Decimal;
     use cosmwasm_std::{
         coin, from_json,
-        testing::{
-            mock_dependencies, mock_env, MockApi, MockQuerier, MockStorage, MOCK_CONTRACT_ADDR,
-        },
-        to_json_binary, Addr, BankMsg, Coin, Coins, CosmosMsg, OwnedDeps, Response, SubMsg,
-        Uint128,
+        testing::{mock_env, MockApi, MockQuerier, MockStorage, MOCK_CONTRACT_ADDR},
+        Addr, BankMsg, Coins, CosmosMsg, OwnedDeps, SubMsg, Uint128,
     };
-    use itertools::Itertools;
+
     use osmosis_std::types::osmosis::tokenfactory::v1beta1::{MsgBurn, MsgMint};
     use osmosis_test_tube::cosmrs::proto::prost::Message;
-    use rstest::rstest;
     use transmuter_math::rebalancing::config::RebalancingConfig;
-
-    #[rstest]
-    #[case(
-        Entrypoint::Exec,
-        SwapToAlloyedConstraint::ExactIn {
-            tokens_in: &[coin(100, "denom1")],
-            token_out_min_amount: Uint128::one(),
-        },
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(MsgMint {
-                sender: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Some(coin(10000u128, "alloyed").into()),
-                mint_to_address: "addr1".to_string()
-            })),
-    )]
-    #[case(
-        Entrypoint::Sudo,
-        SwapToAlloyedConstraint::ExactIn {
-            tokens_in: &[coin(100, "denom1")],
-            token_out_min_amount: Uint128::one(),
-        },
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .set_data(to_json_binary(&SwapExactAmountInResponseData {
-                token_out_amount: Uint128::new(10000u128)
-            }).unwrap())
-            .add_message(MsgMint {
-                sender: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Some(coin(10000u128, "alloyed").into()),
-                mint_to_address: "addr1".to_string()
-            })),
-    )]
-    #[case(
-        Entrypoint::Exec,
-        SwapToAlloyedConstraint::ExactOut {
-            token_in_denom: "denom1",
-            token_in_max_amount: Uint128::new(100),
-            token_out_amount: Uint128::new(10000u128)
-        },
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(MsgMint {
-                sender: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Some(coin(10000u128, "alloyed").into()),
-                mint_to_address: "addr1".to_string()
-            })),
-    )]
-    #[case(
-        Entrypoint::Sudo,
-        SwapToAlloyedConstraint::ExactOut {
-            token_in_denom: "denom1",
-            token_in_max_amount: Uint128::new(100),
-            token_out_amount: Uint128::new(10000u128)
-        },
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .set_data(to_json_binary(&SwapExactAmountOutResponseData {
-                token_in_amount: Uint128::new(100u128)
-            }).unwrap())
-            .add_message(MsgMint {
-                sender: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Some(coin(10000u128, "alloyed").into()),
-                mint_to_address: "addr1".to_string()
-            })),
-    )]
-    fn test_swap_tokens_to_alloyed_asset(
-        #[case] entrypoint: Entrypoint,
-        #[case] constraint: SwapToAlloyedConstraint,
-        #[case] mint_to_address: Addr,
-        #[case] expected_res: Result<Response, ContractError>,
-    ) {
-        let mut deps = mock_dependencies();
-        let transmuter = Transmuter::new();
-        transmuter
-            .alloyed_asset
-            .set_alloyed_denom(&mut deps.storage, &"alloyed".to_string())
-            .unwrap();
-
-        transmuter
-            .alloyed_asset
-            .set_normalization_factor(&mut deps.storage, 100u128.into())
-            .unwrap();
-
-        transmuter
-            .pool
-            .save(
-                &mut deps.storage,
-                &TransmuterPool {
-                    pool_assets: vec![
-                        Asset::new(Uint128::from(1000u128), "denom1", 1u128).unwrap(),
-                        Asset::new(Uint128::from(1000u128), "denom2", 10u128).unwrap(),
-                    ],
-                    asset_groups: BTreeMap::new(),
-                },
-            )
-            .unwrap();
-
-        let res = transmuter.swap_tokens_to_alloyed_asset(
-            entrypoint,
-            constraint,
-            mint_to_address,
-            deps.as_mut(),
-            mock_env(),
-        );
-
-        assert_eq!(res, expected_res);
-    }
-
-    #[rstest]
-    #[case(
-        Entrypoint::Exec,
-        SwapFromAlloyedConstraint::ExactIn {
-            token_out_denom: "denom1",
-            token_out_min_amount: Uint128::from(1u128),
-            token_in_amount: Uint128::from(100u128),
-        },
-        BurnTarget::SenderAccount,
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(MsgBurn {
-                sender: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Some(coin(100u128, "alloyed").into()),
-                burn_from_address: "addr1".to_string()
-            })
-            .add_message(BankMsg::Send {
-                to_address: "addr1".to_string(),
-                amount: vec![coin(1u128, "denom1")]
-            }))
-    )]
-    #[case(
-        Entrypoint::Sudo,
-        SwapFromAlloyedConstraint::ExactIn {
-            token_out_denom: "denom1",
-            token_out_min_amount: Uint128::from(1u128),
-            token_in_amount: Uint128::from(100u128),
-        },
-        BurnTarget::SentFunds,
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(MsgBurn {
-                sender: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Some(coin(100u128, "alloyed").into()),
-                burn_from_address: MOCK_CONTRACT_ADDR.to_string()
-            })
-            .add_message(BankMsg::Send {
-                to_address: "addr1".to_string(),
-                amount: vec![coin(1u128, "denom1")]
-            })
-            .set_data(to_json_binary(&SwapExactAmountInResponseData {
-                token_out_amount: Uint128::from(1u128)
-            }).unwrap()))
-    )]
-    #[case(
-        Entrypoint::Exec,
-        SwapFromAlloyedConstraint::ExactOut {
-            tokens_out: &[coin(1u128, "denom1")],
-            token_in_max_amount: Uint128::from(100u128),
-        },
-        BurnTarget::SenderAccount,
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(MsgBurn {
-                sender: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Some(coin(100u128, "alloyed").into()),
-                burn_from_address: "addr1".to_string()
-            })
-            .add_message(BankMsg::Send {
-                to_address: "addr1".to_string(),
-                amount: vec![coin(1u128, "denom1")]
-            }))
-    )]
-    #[case(
-        Entrypoint::Sudo,
-        SwapFromAlloyedConstraint::ExactOut {
-            tokens_out: &[coin(1u128, "denom1")],
-            token_in_max_amount: Uint128::from(100u128),
-        },
-        BurnTarget::SentFunds,
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(MsgBurn {
-                sender: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Some(coin(100u128, "alloyed").into()),
-                burn_from_address: MOCK_CONTRACT_ADDR.to_string()
-            })
-            .add_message(BankMsg::Send {
-                to_address: "addr1".to_string(),
-                amount: vec![coin(1u128, "denom1")]
-            })
-            .set_data(to_json_binary(&SwapExactAmountOutResponseData {
-                token_in_amount: Uint128::from(100u128)
-            }).unwrap()))
-    )]
-    fn test_swap_alloyed_asset_to_tokens(
-        #[case] entrypoint: Entrypoint,
-        #[case] constraint: SwapFromAlloyedConstraint,
-        #[case] burn_target: BurnTarget,
-        #[case] sender: Addr,
-        #[case] expected_res: Result<Response, ContractError>,
-    ) {
-        let alloyed_holder = match burn_target {
-            BurnTarget::SenderAccount => sender.to_string(),
-            BurnTarget::SentFunds => MOCK_CONTRACT_ADDR.to_string(),
-        };
-
-        let mut deps = cosmwasm_std::testing::mock_dependencies_with_balances(&[(
-            alloyed_holder.as_str(),
-            &[coin(110000000000000u128, "alloyed")],
-        )]);
-
-        let transmuter = Transmuter::new();
-        transmuter
-            .alloyed_asset
-            .set_alloyed_denom(&mut deps.storage, &"alloyed".to_string())
-            .unwrap();
-
-        transmuter
-            .alloyed_asset
-            .set_normalization_factor(&mut deps.storage, 100u128.into())
-            .unwrap();
-
-        transmuter
-            .pool
-            .save(
-                &mut deps.storage,
-                &TransmuterPool {
-                    pool_assets: vec![
-                        Asset::new(Uint128::from(1000000000000u128), "denom1", 1u128).unwrap(),
-                        Asset::new(Uint128::from(1000000000000u128), "denom2", 10u128).unwrap(),
-                    ],
-                    asset_groups: BTreeMap::new(),
-                },
-            )
-            .unwrap();
-
-        let res = transmuter.swap_alloyed_asset_to_tokens(
-            entrypoint,
-            constraint,
-            burn_target,
-            sender,
-            deps.as_mut(),
-            mock_env(),
-        );
-
-        assert_eq!(res, expected_res);
-
-        let pool = transmuter.pool.load(&deps.storage).unwrap();
-
-        for denom in ["denom1", "denom2"] {
-            assert!(pool.has_denom(denom))
-        }
-    }
-
-    #[rstest]
-    #[case(
-        Entrypoint::Sudo,
-        SwapFromAlloyedConstraint::ExactOut {
-            tokens_out: &[coin(1000000000000u128, "denom1")],
-            token_in_max_amount: Uint128::from(100000000000000u128),
-        },
-        vec!["denom1"],
-        vec!["denom1"],
-        BurnTarget::SentFunds,
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(MsgBurn {
-                sender: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Some(coin(100000000000000u128, "alloyed").into()),
-                burn_from_address: MOCK_CONTRACT_ADDR.to_string()
-            })
-            .add_message(BankMsg::Send {
-                to_address: "addr1".to_string(),
-                amount: vec![coin(1000000000000u128, "denom1")]
-            })
-            .set_data(to_json_binary(&SwapExactAmountOutResponseData {
-                token_in_amount: Uint128::from(100000000000000u128)
-            }).unwrap()))
-    )]
-    #[case(
-        Entrypoint::Sudo,
-        SwapFromAlloyedConstraint::ExactIn {
-            token_out_denom: "denom1",
-            token_out_min_amount: 1000000000000u128.into(),
-            token_in_amount: 100000000000000u128.into(),
-        },
-        vec!["denom1"],
-        vec!["denom1"],
-        BurnTarget::SentFunds,
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(MsgBurn {
-                sender: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Some(coin(100000000000000u128, "alloyed").into()),
-                burn_from_address: MOCK_CONTRACT_ADDR.to_string()
-            })
-            .add_message(BankMsg::Send {
-                to_address: "addr1".to_string(),
-                amount: vec![coin(1000000000000u128, "denom1")]
-            })
-            .set_data(to_json_binary(&SwapExactAmountInResponseData {
-                token_out_amount: 1000000000000u128.into(),
-            }).unwrap()))
-    )]
-    #[case(
-        Entrypoint::Exec,
-        SwapFromAlloyedConstraint::ExactIn {
-            token_out_denom: "denom1",
-            token_out_min_amount: 1000000000000u128.into(),
-            token_in_amount: 100000000000000u128.into(),
-        },
-        vec!["denom1"],
-        vec!["denom1"],
-        BurnTarget::SenderAccount,
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(MsgBurn {
-                sender: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Some(coin(100000000000000u128, "alloyed").into()),
-                burn_from_address: "addr1".to_string()
-            })
-            .add_message(BankMsg::Send {
-                to_address: "addr1".to_string(),
-                amount: vec![coin(1000000000000u128, "denom1")]
-            }))
-    )]
-    #[case(
-        Entrypoint::Exec,
-        SwapFromAlloyedConstraint::ExactOut {
-            tokens_out: &[coin(1000000000000u128, "denom1")],
-            token_in_max_amount: Uint128::from(100000000000000u128),
-        },
-        vec!["denom1"],
-        vec!["denom1"],
-        BurnTarget::SenderAccount,
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(MsgBurn {
-                sender: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Some(coin(100000000000000u128, "alloyed").into()),
-                burn_from_address: "addr1".to_string()
-            })
-            .add_message(BankMsg::Send {
-                to_address: "addr1".to_string(),
-                amount: vec![coin(1000000000000u128, "denom1")]
-            }))
-    )]
-    #[case(
-        Entrypoint::Sudo,
-        SwapFromAlloyedConstraint::ExactOut {
-            tokens_out: &[coin(1000000000000u128, "denom1"), coin(1000000000000u128, "denom2")],
-            token_in_max_amount: Uint128::from(110000000000000u128),
-        },
-        vec!["denom1", "denom2"],
-        vec!["denom1", "denom2"],
-        BurnTarget::SentFunds,
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(MsgBurn {
-                sender: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Some(coin(110000000000000u128, "alloyed").into()),
-                burn_from_address: MOCK_CONTRACT_ADDR.to_string()
-            })
-            .add_message(BankMsg::Send {
-                to_address: "addr1".to_string(),
-                amount: vec![coin(1000000000000u128, "denom1"), coin(1000000000000u128, "denom2")]
-            })
-            .set_data(to_json_binary(&SwapExactAmountOutResponseData {
-                token_in_amount: Uint128::from(110000000000000u128),
-            }).unwrap()))
-    )]
-    #[case(
-        Entrypoint::Sudo,
-        SwapFromAlloyedConstraint::ExactOut {
-            tokens_out: &[coin(1000000000000u128, "denom1"), coin(500000000000u128, "denom2")],
-            token_in_max_amount: Uint128::from(105000000000000u128),
-        },
-        vec!["denom1", "denom2"],
-        vec!["denom1"],
-        BurnTarget::SentFunds,
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(MsgBurn {
-                sender: MOCK_CONTRACT_ADDR.to_string(),
-                amount: Some(coin(105000000000000u128, "alloyed").into()),
-                burn_from_address: MOCK_CONTRACT_ADDR.to_string()
-            })
-            .add_message(BankMsg::Send {
-                to_address: "addr1".to_string(),
-                amount: vec![coin(1000000000000u128, "denom1"), coin(500000000000u128, "denom2")],
-            })
-            .set_data(to_json_binary(&SwapExactAmountOutResponseData {
-                token_in_amount: Uint128::from(105000000000000u128),
-            }).unwrap()))
-    )]
-    fn test_swap_alloyed_asset_to_tokens_with_corrupted_assets(
-        #[case] entrypoint: Entrypoint,
-        #[case] constraint: SwapFromAlloyedConstraint,
-        #[case] corrupted_denoms: Vec<&str>,
-        #[case] removed_denoms: Vec<&str>,
-        #[case] burn_target: BurnTarget,
-        #[case] sender: Addr,
-        #[case] expected_res: Result<Response, ContractError>,
-    ) {
-        let alloyed_holder = match burn_target {
-            BurnTarget::SenderAccount => sender.to_string(),
-            BurnTarget::SentFunds => MOCK_CONTRACT_ADDR.to_string(),
-        };
-
-        let mut deps = cosmwasm_std::testing::mock_dependencies_with_balances(&[(
-            alloyed_holder.as_str(),
-            &[coin(210000000000000u128, "alloyed")],
-        )]);
-
-        let transmuter = Transmuter::new();
-        transmuter
-            .alloyed_asset
-            .set_alloyed_denom(&mut deps.storage, &"alloyed".to_string())
-            .unwrap();
-
-        transmuter
-            .alloyed_asset
-            .set_normalization_factor(&mut deps.storage, 100u128.into())
-            .unwrap();
-
-        let mut pool = TransmuterPool {
-            pool_assets: vec![
-                Asset::new(Uint128::from(1000000000000u128), "denom1", 1u128).unwrap(), // 1000000000000 * 100
-                Asset::new(Uint128::from(1000000000000u128), "denom2", 10u128).unwrap(), // 1000000000000 * 10
-                Asset::new(Uint128::from(1000000000000u128), "denom3", 1u128).unwrap(), // 1000000000000 * 100
-            ],
-            asset_groups: BTreeMap::new(),
-        };
-
-        let all_denoms = pool
-            .clone()
-            .pool_assets
-            .into_iter()
-            .map(|asset| asset.denom().to_string())
-            .collect::<Vec<String>>();
-
-        for denom in corrupted_denoms {
-            pool.mark_corrupted_asset(denom).unwrap();
-        }
-
-        transmuter.pool.save(&mut deps.storage, &pool).unwrap();
-
-        for denom in all_denoms.clone() {
-            transmuter
-                .rebalancer
-                .add_config(
-                    &mut deps.storage,
-                    Scope::denom(denom.as_str()),
-                    RebalancingConfig::limit_only(Decimal::percent(100)).unwrap(),
-                )
-                .unwrap();
-        }
-
-        let res = transmuter.swap_alloyed_asset_to_tokens(
-            entrypoint,
-            constraint,
-            burn_target,
-            sender,
-            deps.as_mut(),
-            mock_env(),
-        );
-
-        assert_eq!(res, expected_res);
-
-        // all drained denoms that are corrupted should not be in the pool
-        let pool = transmuter.pool.load(&deps.storage).unwrap();
-
-        for denom in all_denoms {
-            if removed_denoms.contains(&denom.as_str()) {
-                assert!(
-                    !pool.has_denom(denom.as_str()),
-                    "must not contain {} since it's corrupted and drained",
-                    denom
-                );
-
-                // limiters should be removed
-                assert!(
-                    transmuter
-                        .rebalancer
-                        .get_config_by_scope(&deps.storage, &Scope::denom(denom.as_str()))
-                        .unwrap()
-                        .is_none(),
-                    "must not contain limiter for {} since it's corrupted and drained",
-                    denom
-                );
-            } else {
-                assert!(
-                    pool.has_denom(denom.as_str()),
-                    "must contain {} since it's not corrupted or not drained",
-                    denom
-                );
-
-                // limiters should not be removed
-                assert!(
-                    transmuter
-                        .rebalancer
-                        .get_config_by_scope(&deps.storage, &Scope::denom(denom.as_str()))
-                        .unwrap()
-                        .is_some(),
-                    "must contain limiter for {} since it's not corrupted or not drained",
-                    denom
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn test_swap_non_alloyed_exact_amount_in_with_corrupted_assets() {
-        let mut deps = mock_dependencies();
-        let transmuter = Transmuter::new();
-        transmuter
-            .alloyed_asset
-            .set_alloyed_denom(&mut deps.storage, &"alloyed".to_string())
-            .unwrap();
-
-        transmuter
-            .alloyed_asset
-            .set_normalization_factor(&mut deps.storage, 100u128.into())
-            .unwrap();
-
-        let mut pool = TransmuterPool {
-            pool_assets: vec![
-                Asset::new(Uint128::from(1000000000000u128), "denom1", 1u128).unwrap(), // 1000000000000 * 100
-                Asset::new(Uint128::from(1000000000000u128), "denom2", 10u128).unwrap(), // 1000000000000 * 10
-                Asset::new(Uint128::from(1000000000000u128), "denom3", 1u128).unwrap(), // 1000000000000 * 100
-            ],
-            asset_groups: BTreeMap::new(),
-        };
-
-        let all_denoms = pool
-            .clone()
-            .pool_assets
-            .into_iter()
-            .map(|asset| asset.denom().to_string())
-            .collect::<Vec<_>>();
-
-        pool.mark_corrupted_asset("denom1").unwrap();
-
-        transmuter.pool.save(&mut deps.storage, &pool).unwrap();
-
-        for denom in all_denoms.clone() {
-            transmuter
-                .rebalancer
-                .add_config(
-                    &mut deps.storage,
-                    Scope::denom(denom.as_str()),
-                    RebalancingConfig::limit_only(Decimal::percent(100)).unwrap(),
-                )
-                .unwrap();
-        }
-
-        transmuter
-            .swap_non_alloyed_exact_amount_in(
-                coin(1000000000000, "denom3"),
-                "denom1",
-                1000000000000u128.into(),
-                deps.api.addr_make("sender"),
-                deps.as_mut(),
-            )
-            .unwrap();
-
-        // all drained denoms that are corrupted should not be in the pool
-        let pool = transmuter.pool.load(&deps.storage).unwrap();
-
-        let denoms = pool
-            .pool_assets
-            .into_iter()
-            .map(|a| a.denom().to_string())
-            .collect_vec();
-
-        assert_eq!(denoms, vec!["denom2", "denom3"]);
-
-        let limiter_denoms = transmuter
-            .rebalancer
-            .list_configs(&deps.storage)
-            .unwrap()
-            .into_iter()
-            .map(|(denom, _)| denom)
-            .unique()
-            .collect_vec();
-
-        assert_eq!(
-            limiter_denoms,
-            vec![Scope::denom("denom2").key(), Scope::denom("denom3").key()]
-        );
-    }
-
-    #[test]
-    fn test_swap_non_alloyed_exact_amount_out_with_corrupted_assets() {
-        let mut deps = mock_dependencies();
-        let transmuter = Transmuter::new();
-        transmuter
-            .alloyed_asset
-            .set_alloyed_denom(&mut deps.storage, &"alloyed".to_string())
-            .unwrap();
-
-        transmuter
-            .alloyed_asset
-            .set_normalization_factor(&mut deps.storage, 100u128.into())
-            .unwrap();
-
-        let mut pool = TransmuterPool {
-            pool_assets: vec![
-                Asset::new(Uint128::from(1000000000000u128), "denom1", 1u128).unwrap(), // 1000000000000 * 100
-                Asset::new(Uint128::from(1000000000000u128), "denom2", 10u128).unwrap(), // 1000000000000 * 10
-                Asset::new(Uint128::from(1000000000000u128), "denom3", 1u128).unwrap(), // 1000000000000 * 100
-            ],
-            asset_groups: BTreeMap::new(),
-        };
-
-        let all_denoms = pool
-            .clone()
-            .pool_assets
-            .into_iter()
-            .map(|asset| asset.denom().to_string())
-            .collect::<Vec<_>>();
-
-        pool.mark_corrupted_asset("denom1").unwrap();
-
-        transmuter.pool.save(&mut deps.storage, &pool).unwrap();
-
-        for denom in all_denoms.clone() {
-            transmuter
-                .rebalancer
-                .add_config(
-                    &mut deps.storage,
-                    Scope::denom(denom.as_str()),
-                    RebalancingConfig::limit_only(Decimal::percent(100)).unwrap(),
-                )
-                .unwrap();
-        }
-
-        transmuter
-            .swap_non_alloyed_exact_amount_out(
-                "denom3",
-                1000000000000u128.into(),
-                coin(1000000000000, "denom1"),
-                deps.api.addr_make("sender"),
-                deps.as_mut(),
-            )
-            .unwrap();
-
-        // all drained denoms that are corrupted should not be in the pool
-        let pool = transmuter.pool.load(&deps.storage).unwrap();
-
-        let denoms = pool
-            .pool_assets
-            .into_iter()
-            .map(|a| a.denom().to_string())
-            .collect_vec();
-
-        assert_eq!(denoms, vec!["denom2", "denom3"]);
-
-        let limiter_denoms = transmuter
-            .rebalancer
-            .list_configs(&deps.storage)
-            .unwrap()
-            .into_iter()
-            .map(|(denom, _)| denom)
-            .unique()
-            .collect_vec();
-
-        assert_eq!(
-            limiter_denoms,
-            vec![Scope::denom("denom2").key(), Scope::denom("denom3").key()]
-        );
-    }
-
-    #[rstest]
-    #[case(
-        coin(100u128, "denom1"),
-        "denom2",
-        1000u128,
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(BankMsg::Send {
-                to_address: "addr1".to_string(),
-                amount: vec![coin(1000u128, "denom2")]
-            })
-            .set_data(to_json_binary(&SwapExactAmountInResponseData {
-                token_out_amount: Uint128::from(1000u128)
-            }).unwrap()))
-    )]
-    #[case(
-        coin(100u128, "denom2"),
-        "denom1",
-        10u128,
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(BankMsg::Send {
-                to_address: "addr1".to_string(),
-                amount: vec![coin(10u128, "denom1")]
-            })
-            .set_data(to_json_binary(&SwapExactAmountInResponseData {
-                token_out_amount: Uint128::from(10u128)
-            }).unwrap()))
-    )]
-    #[case(
-        coin(100u128, "denom2"),
-        "denom1",
-        100u128,
-        Addr::unchecked("addr1"),
-        Err(ContractError::InsufficientTokenOut {
-            min_required: 100u128.into(),
-            amount_out: 10u128.into()
-        })
-    )]
-    #[case(
-        coin(100000000001u128, "denom1"),
-        "denom2",
-        1000000000010u128,
-        Addr::unchecked("addr1"),
-        Err(ContractError::InsufficientPoolAsset {
-            required: coin(1000000000010u128, "denom2"),
-            available: coin(1000000000000u128, "denom2"),
-        })
-    )]
-    fn test_swap_non_alloyed_exact_amount_in(
-        #[case] token_in: Coin,
-        #[case] token_out_denom: &str,
-        #[case] token_out_min_amount: u128,
-        #[case] sender: Addr,
-        #[case] expected_res: Result<Response, ContractError>,
-    ) {
-        let mut deps = cosmwasm_std::testing::mock_dependencies_with_balances(&[(
-            sender.to_string().as_str(),
-            &[coin(2000000000000u128, "alloyed")],
-        )]);
-
-        let transmuter = Transmuter::new();
-        transmuter
-            .alloyed_asset
-            .set_alloyed_denom(&mut deps.storage, &"alloyed".to_string())
-            .unwrap();
-
-        transmuter
-            .alloyed_asset
-            .set_normalization_factor(&mut deps.storage, 100u128.into())
-            .unwrap();
-
-        transmuter
-            .pool
-            .save(
-                &mut deps.storage,
-                &TransmuterPool {
-                    pool_assets: vec![
-                        Asset::new(Uint128::from(1000000000000u128), "denom1", 1u128).unwrap(),
-                        Asset::new(Uint128::from(1000000000000u128), "denom2", 10u128).unwrap(),
-                    ],
-                    asset_groups: BTreeMap::new(),
-                },
-            )
-            .unwrap();
-
-        let res = transmuter.swap_non_alloyed_exact_amount_in(
-            token_in.clone(),
-            token_out_denom,
-            token_out_min_amount.into(),
-            sender,
-            deps.as_mut(),
-        );
-
-        assert_eq!(res, expected_res);
-    }
-
-    #[rstest]
-    #[case(
-        "denom1",
-        100u128,
-        coin(1000u128, "denom2"),
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(BankMsg::Send {
-                to_address: "addr1".to_string(),
-                amount: vec![coin(1000u128, "denom2")]
-            })
-            .set_data(to_json_binary(&SwapExactAmountOutResponseData {
-                token_in_amount: 100u128.into()
-            }).unwrap()))
-    )]
-    #[case(
-        "denom2",
-        100u128,
-        coin(10u128, "denom1"),
-        Addr::unchecked("addr1"),
-        Ok(Response::new()
-            .add_message(BankMsg::Send {
-                to_address: "addr1".to_string(),
-                amount: vec![coin(10u128, "denom1")]
-            })
-            .set_data(to_json_binary(&SwapExactAmountOutResponseData {
-                token_in_amount: 100u128.into()
-            }).unwrap()))
-    )]
-    #[case(
-        "denom2",
-        100u128,
-        coin(100u128, "denom1"),
-        Addr::unchecked("addr1"),
-        Err(ContractError::ExcessiveRequiredTokenIn {
-            limit: 100u128.into(),
-            required: 1000u128.into()
-        })
-    )]
-    #[case(
-        "denom1",
-        100000000001u128,
-        coin(1000000000010u128, "denom2"),
-        Addr::unchecked("addr1"),
-        Err(ContractError::InsufficientPoolAsset {
-            required: coin(1000000000010u128, "denom2"),
-            available: coin(1000000000000u128, "denom2"),
-        })
-    )]
-    fn test_swap_non_alloyed_exact_amount_out(
-        #[case] token_in_denom: &str,
-        #[case] token_in_max_amount: u128,
-        #[case] token_out: Coin,
-        #[case] sender: Addr,
-        #[case] expected_res: Result<Response, ContractError>,
-    ) {
-        let mut deps = cosmwasm_std::testing::mock_dependencies_with_balances(&[(
-            sender.to_string().as_str(),
-            &[coin(2000000000000u128, "alloyed")],
-        )]);
-
-        let transmuter = Transmuter::new();
-        transmuter
-            .alloyed_asset
-            .set_alloyed_denom(&mut deps.storage, &"alloyed".to_string())
-            .unwrap();
-
-        transmuter
-            .alloyed_asset
-            .set_normalization_factor(&mut deps.storage, 100u128.into())
-            .unwrap();
-
-        transmuter
-            .pool
-            .save(
-                &mut deps.storage,
-                &TransmuterPool {
-                    pool_assets: vec![
-                        Asset::new(Uint128::from(1000000000000u128), "denom1", 1u128).unwrap(),
-                        Asset::new(Uint128::from(1000000000000u128), "denom2", 10u128).unwrap(),
-                    ],
-                    asset_groups: BTreeMap::new(),
-                },
-            )
-            .unwrap();
-
-        let res = transmuter.swap_non_alloyed_exact_amount_out(
-            token_in_denom,
-            token_in_max_amount.into(),
-            token_out,
-            sender,
-            deps.as_mut(),
-        );
-
-        assert_eq!(res, expected_res);
-    }
-
-    #[rstest]
-    #[case::empty(
-        BTreeMap::new(),
-        BTreeMap::new(),
-        BTreeMap::new(),
-        BTreeMap::new(),
-        vec![],
-    )]
-    #[case::no_prev_asset_weights(
-        BTreeMap::new(),
-        BTreeMap::new(),
-        BTreeMap::from([
-            ("eth.axl".to_string(), Decimal::percent(40)),
-            ("eth.wh".to_string(), Decimal::percent(40)),
-            ("wsteth.axl".to_string(), Decimal::percent(20)),
-        ]),
-        BTreeMap::new(),
-        vec![
-            (Scope::denom("eth.axl"), (Decimal::zero(), Decimal::percent(40))),
-            (Scope::denom("eth.wh"), (Decimal::zero(), Decimal::percent(40))),
-            (Scope::denom("wsteth.axl"), (Decimal::zero(), Decimal::percent(20))),
-        ],
-    )]
-    #[case::no_updated_asset_weights(
-        BTreeMap::from([
-            ("eth.axl".to_string(), Decimal::percent(20)),
-            ("eth.wh".to_string(), Decimal::percent(60)),
-            ("wsteth.axl".to_string(), Decimal::percent(20)),
-        ]),
-        BTreeMap::new(),
-        BTreeMap::new(),
-        BTreeMap::new(),
-        vec![
-            (Scope::denom("eth.axl"), (Decimal::percent(20), Decimal::zero())),
-            (Scope::denom("eth.wh"), (Decimal::percent(60), Decimal::zero())),
-            (Scope::denom("wsteth.axl"), (Decimal::percent(20), Decimal::zero())),
-        ],
-    )]
-    #[case(
-        BTreeMap::from([
-            ("eth.axl".to_string(), Decimal::percent(20)),
-            ("eth.wh".to_string(), Decimal::percent(60)),
-            ("wsteth.axl".to_string(), Decimal::percent(20)),
-        ]),
-        BTreeMap::from([
-            ("axelar".to_string(), Decimal::percent(40)),
-            ("wormhole".to_string(), Decimal::percent(60)),
-        ]),
-        BTreeMap::from([
-            ("eth.axl".to_string(), Decimal::percent(40)),
-            ("eth.wh".to_string(), Decimal::percent(40)),
-            ("wsteth.axl".to_string(), Decimal::percent(20)),
-        ]),
-        BTreeMap::from([
-            ("axelar".to_string(), Decimal::percent(60)),
-            ("wormhole".to_string(), Decimal::percent(40)),
-        ]),
-        vec![
-            (Scope::denom("eth.axl"), (Decimal::percent(20), Decimal::percent(40))),
-            (Scope::denom("eth.wh"), (Decimal::percent(60), Decimal::percent(40))),
-            (Scope::denom("wsteth.axl"), (Decimal::percent(20), Decimal::percent(20))),
-            (Scope::asset_group("axelar"), (Decimal::percent(40), Decimal::percent(60))),
-            (Scope::asset_group("wormhole"), (Decimal::percent(60), Decimal::percent(40))),
-        ],
-    )]
-    fn test_construct_scope_value_pairs(
-        #[case] prev_asset_weights: BTreeMap<String, Decimal>,
-        #[case] prev_asset_group_weights: BTreeMap<String, Decimal>,
-        #[case] updated_asset_weights: BTreeMap<String, Decimal>,
-        #[case] updated_asset_group_weights: BTreeMap<String, Decimal>,
-        #[case] expected_scope_value_pairs: Vec<(Scope, (Decimal, Decimal))>,
-    ) {
-        let mut scope_value_pairs = construct_scope_value_pairs(
-            prev_asset_weights,
-            updated_asset_weights,
-            prev_asset_group_weights,
-            updated_asset_group_weights,
-        )
-        .unwrap();
-
-        // assert by disregard order
-        scope_value_pairs.sort_by_key(|(scope, _)| scope.key());
-        let mut expected_scope_value_pairs = expected_scope_value_pairs;
-        expected_scope_value_pairs.sort_by_key(|(scope, _)| scope.key());
-
-        assert_eq!(scope_value_pairs, expected_scope_value_pairs);
-    }
-
-    #[test]
-    fn test_clean_up_drained_corrupted_assets_group() {
-        let sender = Addr::unchecked("addr1");
-        let mut deps = cosmwasm_std::testing::mock_dependencies_with_balances(&[(
-            sender.to_string().as_str(),
-            &[coin(2000000000000u128, "alloyed")],
-        )]);
-
-        let transmuter = Transmuter::new();
-        transmuter
-            .alloyed_asset
-            .set_alloyed_denom(&mut deps.storage, &"alloyed".to_string())
-            .unwrap();
-
-        transmuter
-            .alloyed_asset
-            .set_normalization_factor(&mut deps.storage, 100u128.into())
-            .unwrap();
-
-        let init_pool = TransmuterPool {
-            pool_assets: vec![
-                Asset::new(Uint128::from(1000000000000u128), "denom1", 1u128).unwrap(),
-                Asset::new(Uint128::from(1000000000000u128), "denom2", 10u128).unwrap(),
-                Asset::new(Uint128::from(1000000000000u128), "denom3", 100u128).unwrap(),
-            ],
-            asset_groups: BTreeMap::from([(
-                "group1".to_string(),
-                AssetGroup::new(vec!["denom2".to_string(), "denom3".to_string()])
-                    .mark_as_corrupted()
-                    .clone(),
-            )]),
-        };
-        transmuter.pool.save(&mut deps.storage, &init_pool).unwrap();
-
-        // Add a rebalancing config for the group
-        transmuter
-            .rebalancer
-            .add_config(
-                &mut deps.storage,
-                Scope::asset_group("group1"),
-                RebalancingConfig::limit_only(Decimal::percent(60)).unwrap(),
-            )
-            .unwrap();
-
-        let mut pool = transmuter.pool.load(&deps.storage).unwrap();
-        let res = transmuter.clean_up_drained_corrupted_assets(&mut deps.storage, &mut pool);
-        assert_eq!(res, Ok(()));
-
-        pool = transmuter.pool.load(&deps.storage).unwrap();
-        assert_eq!(pool, init_pool);
-
-        pool.exit_pool(&[coin(1000000000000u128, "denom2")])
-            .unwrap();
-        transmuter.pool.save(&mut deps.storage, &pool).unwrap();
-
-        let res = transmuter.clean_up_drained_corrupted_assets(&mut deps.storage, &mut pool);
-        assert_eq!(res, Ok(()));
-
-        let expected_pool = TransmuterPool {
-            pool_assets: vec![
-                Asset::new(Uint128::from(1000000000000u128), "denom1", 1u128).unwrap(),
-                Asset::new(Uint128::from(1000000000000u128), "denom3", 100u128).unwrap(),
-            ],
-            asset_groups: BTreeMap::from([(
-                "group1".to_string(),
-                AssetGroup::new(vec!["denom3".to_string()])
-                    .mark_as_corrupted()
-                    .clone(),
-            )]),
-        };
-        assert_eq!(pool, expected_pool);
-
-        // Check that the rebalancing config for group1 is still exists
-        let rebalancing_configs = transmuter
-            .rebalancer
-            .get_config_by_scope(&deps.storage, &Scope::asset_group("group1"))
-            .unwrap();
-        assert!(rebalancing_configs.is_some());
-
-        // Save the updated pool
-        transmuter.pool.save(&mut deps.storage, &pool).unwrap();
-
-        pool.exit_pool(&[coin(1000000000000u128, "denom3")])
-            .unwrap();
-
-        let res = transmuter.clean_up_drained_corrupted_assets(&mut deps.storage, &mut pool);
-        assert_eq!(res, Ok(()));
-
-        let expected_pool = TransmuterPool {
-            pool_assets: vec![
-                Asset::new(Uint128::from(1000000000000u128), "denom1", 1u128).unwrap(),
-            ],
-            asset_groups: BTreeMap::new(),
-        };
-        assert_eq!(pool, expected_pool);
-
-        // Check that the rebalancing config for group1 is removed
-        let rebalancing_configs = transmuter
-            .rebalancer
-            .get_config_by_scope(&deps.storage, &Scope::asset_group("group1"))
-            .unwrap();
-        assert_eq!(rebalancing_configs, None);
-    }
-
-    #[test]
-    fn test_clean_up_drained_corrupted_assets_group_not_corrupted() {
-        let mut deps = mock_dependencies();
-        let transmuter = Transmuter::new();
-
-        // Initialize the pool with non-corrupted assets and groups
-        let init_pool = TransmuterPool {
-            pool_assets: vec![
-                Asset::new(Uint128::from(1000000000000u128), "denom1", 1u128).unwrap(),
-                Asset::new(Uint128::from(1000000000000u128), "denom2", 10u128).unwrap(),
-                Asset::new(Uint128::from(1000000000000u128), "denom3", 100u128).unwrap(),
-            ],
-            asset_groups: BTreeMap::from([(
-                "group1".to_string(),
-                AssetGroup::new(vec!["denom2".to_string(), "denom3".to_string()]),
-            )]),
-        };
-
-        transmuter.pool.save(&mut deps.storage, &init_pool).unwrap();
-
-        // Register a limiter for the group
-        transmuter
-            .rebalancer
-            .add_config(
-                &mut deps.storage,
-                Scope::asset_group("group1"),
-                RebalancingConfig::limit_only(Decimal::one()).unwrap(),
-            )
-            .unwrap();
-
-        let mut pool = transmuter.pool.load(&deps.storage).unwrap();
-        assert_eq!(pool, init_pool);
-
-        // Drain denom2 from the pool
-        pool.exit_pool(&[coin(1000000000000u128, "denom2")])
-            .unwrap();
-        transmuter.pool.save(&mut deps.storage, &pool).unwrap();
-
-        let res = transmuter.clean_up_drained_corrupted_assets(&mut deps.storage, &mut pool);
-        assert_eq!(res, Ok(()));
-
-        // Check that the pool remains unchanged
-        let expected_pool = TransmuterPool {
-            pool_assets: vec![
-                Asset::new(Uint128::from(1000000000000u128), "denom1", 1u128).unwrap(),
-                Asset::new(Uint128::zero(), "denom2", 10u128).unwrap(),
-                Asset::new(Uint128::from(1000000000000u128), "denom3", 100u128).unwrap(),
-            ],
-            asset_groups: BTreeMap::from([(
-                "group1".to_string(),
-                AssetGroup::new(vec!["denom2".to_string(), "denom3".to_string()]),
-            )]),
-        };
-        assert_eq!(pool, expected_pool);
-
-        // Check that the rebalancing config for group1 is still registered
-        let rebalancing_configs = transmuter
-            .rebalancer
-            .get_config_by_scope(&deps.storage, &Scope::asset_group("group1"))
-            .unwrap();
-        assert!(rebalancing_configs.is_some());
-
-        // Save the updated pool
-        transmuter.pool.save(&mut deps.storage, &pool).unwrap();
-
-        // Drain denom3 from the pool
-        pool.exit_pool(&[coin(1000000000000u128, "denom3")])
-            .unwrap();
-
-        let res = transmuter.clean_up_drained_corrupted_assets(&mut deps.storage, &mut pool);
-        assert_eq!(res, Ok(()));
-
-        // Check that the pool remains unchanged except for the drained assets
-        let expected_pool = TransmuterPool {
-            pool_assets: vec![
-                Asset::new(Uint128::from(1000000000000u128), "denom1", 1u128).unwrap(),
-                Asset::new(Uint128::zero(), "denom2", 10u128).unwrap(),
-                Asset::new(Uint128::zero(), "denom3", 100u128).unwrap(),
-            ],
-            asset_groups: BTreeMap::from([(
-                "group1".to_string(),
-                AssetGroup::new(vec!["denom2".to_string(), "denom3".to_string()]),
-            )]),
-        };
-        assert_eq!(pool, expected_pool);
-
-        // Check that the rebalancing config for group1 is still registered
-        let rebalancing_configs = transmuter
-            .rebalancer
-            .get_config_by_scope(&deps.storage, &Scope::asset_group("group1"))
-            .unwrap();
-        assert!(rebalancing_configs.is_some());
-    }
 
     fn setup_fee_deduction_test() -> (Addr, OwnedDeps<MockStorage, MockApi, MockQuerier>) {
         let sender = Addr::unchecked("sender");
